+++ public/manifest.json
@@ -34,7 +34,7 @@
     {
       "matches": ["<all_urls>"],
       "run_at": "document_idle",
-      "js": ["contentScript.js", "popup.js", "minifyHtml.js"]
+      "js": ["contentScript.js", "popup.js", "minifyHtml.js", "flowHandler.js"]
     }
   ]
 }
